move compile detail graph to component
